Remove custom pinentry
This is intended to be temporary until I completely finish migrating to wayland
and replacing things like rofi.  In the meantime, if I need to do rebuilds on a
tty, the custom pinentry breaks since it requires X.

When I resurrect this, I'll make sure the replacement works for both wayland and
a tty.