[emacs] Extract configuration to a custom emacs package
This PR takes my emacs configuration and migrates it to a custom elisp package
called `emacs-rc`.  It also fixes a few issues that were discovered thanks to
compilation errors while migrating--deprecated APIs, etc.

I chose to un-modularize my configuration as part of this because:

1) I was importing everything from a central place, which the compiler --
operating on one file at a time -- couldn't possibly be aware of.  So I'd need
to at least restructure things so that a file could require it's dependencies
rather than depending on being imported in the correct order.

2) The modular structure I had didn't really make much sense and it's not any
harder to find things with a flat single file for now.

3) I have an old attempt at building an emacs-rc that was a ground-up rewrite
and was done as a single file.  Being able to diff with that file to pull in
things I want will make my life much easier.

Probably I'll re-modularize the file again in the "eventual future".