[emacs] Don't overuse `:mode` when I can use `:demand`
I think in general `:mode` is more performant at emacs startup time since it
defers loading of the entire dependency.  But there are caveats: it introduces a
perf hit on the first time the dependency is needed, and I must ensure my
regexes in the `:mode` sections stay synced with what the dependency would have
used internally.  Since I use server mode, it's not all that important to me to
optimize emacs startup time.  So this change just uses `:demand` so I don't have
to deal with synchronizing regexes.