[detachedLuksWithNixopsKeys] Fix auto-dependency & add stop script
I used to have this unit wanted by zfs.target and never removed that when making
this more generic.  Now that I use it for drives that aren't part of
zfs.target (and aren't always plugged in) it's causing systemd failures to show
up a lot.  This change removes that line to fix those errors.

I also added a convenience stop script so I can stop the unit to close the
encrypted FS.