[sources] Upgrade sources
This upgrade includes a change to nixops that fixes using `nixpkgs.overlays`, so
this change also moves to `nixpkgs.overlays` instead of using `nixpkgs-overlays`
in the `NIX_PATH`.  I'll need to follow this up with some cleanup to the
overlays structure in general now that I can put overlays in more natural places.