[emacs] Fix some issues while compiling emacs-rc
1. Nixpkgs currently is double-compiling packages when native compilation is on.
First, it's doing a byte-compile, then it's doing a native compile, which is
redundant as the native compile also produces byte compiled outputs.  In this PR
I set the `buildPhase` to `null` to disable the byte compile.

2. In the native compile, emacs has a bug that causes it to try to load files
after compiling, which causes some weird but innocuous errors and slows down the
compile.  This PR hacks a temporary fix in.  Eventually I'll put a fix in to
emacs itself.