[wireguard] Add hack to fix wireguard not connecting on boot
Occasionally, wireguard fails on boot.  The cause is:

1. dhcpcd reports it's up before an IP address actually is received, which
causes `network-online.target` to be reached even though there isn't any online
network.

2. The wireguard units run after `network-online.target`

3. If wireguard receives a "permanent" dns failure -- as it would if there's no
resolver (yet) -- then rather than retrying the dns resolution, it crashes.

4. The systemd units generated by nixos don't restart on failure

This PR implements a hack fix to make the dhcpcd unit wait until it actually has
an ipv4 lease to report it's up and move on with `network-online.target`.  This
is not a good fix to the wireguard issues -- the nixos units should be more
resilient -- but until something lands upstream this is a safe and minimally
invasive fix.