[emacs] Set up emacs-hledger
This commit also redesigns how system dependencies are exposed to `emacs-rc`.

Previously, we used custom configuration to pass paths to specific executables
in to emacs config files, and then would attempt to thread those paths to the
correct underlying configs.  This mostly worked, but it's clunky, and it does
not work in cases where the underlying libraries have the program name hardcoded
and no way to pass in the path, assuming that the path can be resolved from the
exec-path (such as is the case with `emacs-hledger`).

Now, we simply have an overlay which creates a wrapper around emacs that sets
the `$PATH` env var to include all the needed dependencies instead.